+++ docs/src/version_control.f90
@@ -2,6 +2,6 @@ module version_control
     !! Keeps track of the current version of the program
     implicit none
 
-    character(len=64) :: version = "0.2.0"
+    character(len=64) :: version = "0.2.1"
 
 end module